add flag for PUT, restructured code for generate
